# specs/001-news-analysis/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: News Analysis API
  version: 0.1.0
  description: 'Contract for analyzing a news URL and returning an operational reliability
    index with traceable criterion evidence. Scores are not probabilities of truth
    or falsity.

    '
paths:
  /analyses:
    post:
      summary: Analyze one news URL
      operationId: createAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
      responses:
        '200':
          description: Analysis completed or completed with unavailable criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Invalid URL or unsupported URL scheme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: URL blocked for security
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: External fetch, fact-checking, or model execution failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /analyses/{analysis_id}:
    get:
      summary: Retrieve an audit record by analysis id
      operationId: getAnalysis
      parameters:
      - name: analysis_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Stored audit record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '404':
          description: Analysis not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    AnalysisRequest:
      additionalProperties: false
      properties:
        url:
          title: Url
          type: string
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: User Id
        claim:
          anyOf:
          - maxLength: 500
            minLength: 3
            type: string
          - type: 'null'
          default: null
          title: Claim
      required:
      - url
      title: AnalysisRequest
      type: object
    AnalysisResponse:
      type: object
      required:
      - id
      - status
      - input
      - criteria
      - final
      - pipeline_version
      - limitations
      - created_at
      properties:
        id:
          type: string
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        input:
          type: object
          required:
          - url
          properties:
            url:
              type: string
              format: uri
            claim:
              type: string
              description: Optional user-selected claim; evaluated independently of
                source reputation.
          additionalProperties: false
        article:
          $ref: '#/components/schemas/Article'
        criteria:
          $ref: '#/components/schemas/Criteria'
        final:
          $ref: '#/components/schemas/FinalScore'
        pipeline_version:
          $ref: '#/components/schemas/PipelineVersion'
        limitations:
          type: array
          items:
            type: string
        error:
          $ref: '#/components/schemas/ErrorInfo'
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
      additionalProperties: false
    AnalysisStatus:
      type: string
      enum:
      - SUCCESS
      - INVALID_URL
      - BLOCKED_INTERNAL_URL
      - NEWS_FETCH_FAILED
      - ARTICLE_EXTRACTION_FAILED
      - FACT_CHECK_API_ERROR
      - WRITING_MODEL_ERROR
      - CRITERION_UNAVAILABLE
      - RATE_LIMITED
    Article:
      type: object
      required:
      - original_url
      - content_hash
      - extracted_character_count
      properties:
        original_url:
          type: string
          format: uri
        final_url:
          type: string
          format: uri
        canonical_url:
          type: string
          format: uri
        publisher:
          type: string
        title:
          type: string
        subtitle:
          type: string
        author:
          type: string
        published_at:
          type: string
        content_hash:
          type: string
        extracted_character_count:
          type: integer
          minimum: 1000
      additionalProperties: false
    Criteria:
      type: object
      required:
      - verifiable_facts
      - writing_style
      - factual_claims
      properties:
        verifiable_facts:
          $ref: '#/components/schemas/FactCheckCriterion'
        writing_style:
          $ref: '#/components/schemas/WritingStyleCriterion'
        factual_claims:
          $ref: '#/components/schemas/ReservedCriterion'
      additionalProperties: false
    FactCheckCriterion:
      additionalProperties: false
      properties:
        available:
          title: Available
          type: boolean
        status:
          $ref: '#/components/schemas/CriterionStatus'
        score:
          anyOf:
          - maximum: 1
            minimum: 0
            type: number
          - type: 'null'
          default: null
          title: Score
        name:
          default: Checagem de fatos verificáveis
          title: Name
          type: string
        method:
          default: google_fact_check_claim_reviews
          title: Method
          type: string
        target_claim:
          anyOf:
          - type: string
          - type: 'null'
          default: null
          title: Target Claim
        claim_origin:
          default: article_title_or_excerpt
          title: Claim Origin
          type: string
        scope:
          default: Uma afirmação selecionada; não verifica todos os fatos da notícia
            nem a reputação da fonte.
          title: Scope
          type: string
        formula:
          default: F = mean(publisher means of unique applicable mapped reviews)
          title: Formula
          type: string
        limitation:
          default: Google retrieves published reviews; matching is lexical, the score
            is project-defined, and absence of a review is not a verdict.
          title: Limitation
          type: string
        scored_reviews_count:
          default: 0
          title: Scored Reviews Count
          type: integer
        publishers_count:
          default: 0
          title: Publishers Count
          type: integer
        publisher_scores:
          additionalProperties:
            type: number
          title: Publisher Scores
          type: object
        conflicting_verdicts:
          default: false
          title: Conflicting Verdicts
          type: boolean
        search_attempts:
          items:
            additionalProperties: true
            type: object
          title: Search Attempts
          type: array
        search_truncated:
          default: false
          title: Search Truncated
          type: boolean
        query:
          default: ''
          title: Query
          type: string
        reviews_count:
          minimum: 0
          title: Reviews Count
          type: integer
        applicable_reviews_count:
          minimum: 0
          title: Applicable Reviews Count
          type: integer
        reviews:
          items:
            $ref: '#/components/schemas/FactCheckReview'
          title: Reviews
          type: array
        intended_weight:
          default: 0.6
          title: Intended Weight
          type: number
        effective_weight:
          anyOf:
          - type: number
          - type: 'null'
          default: null
          title: Effective Weight
        contribution:
          anyOf:
          - type: number
          - type: 'null'
          default: null
          title: Contribution
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorInfo'
          - type: 'null'
          default: null
      required:
      - available
      - status
      - reviews_count
      - applicable_reviews_count
      - reviews
      title: FactCheckCriterionResult
      type: object
...
    WritingStyleCriterion:
      type: object
      required:
      - available
      - status
      - score
      - model
      - model_version
      - segments_analyzed
      - segments
      - qualitative_state
      - limitation
      - intended_weight
      properties:
        available:
          type: boolean
        status:
          type: string
          enum:
          - EXECUTED
          - UNAVAILABLE
          - ERROR
        score:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 1
        model:
          type: string
          const: vzani/portuguese-fake-news-classifier-bertimbau-combined
        model_version:
          type: string
          description: Pinned BERTimbau weights revision used for inference.
          const: 86971e56e7f5ad781cf56673df73a57375455793
        prediction:
          $ref: '#/components/schemas/WritingPrediction'
        segments_analyzed:
          type: integer
          minimum: 0
          description: Number of non-overlapping token windows classified; zero on
            criterion error.
        segments:
          type: array
          items:
            $ref: '#/components/schemas/WritingSegment'
        qualitative_state:
          type:
          - string
          - 'null'
          enum:
          - Sinal de escrita suspeito
          - null
        limitation:
          type: string
        intended_weight:
          type: number
          const: 0.4
        effective_weight:
          type:
          - number
          - 'null'
        contribution:
          type:
          - number
          - 'null'
        error:
          $ref: '#/components/schemas/ErrorInfo'
      additionalProperties: false
    WritingPrediction:
      type: object
      description: Character-weighted aggregate; True for score >= 0.5, otherwise
        Fake. Confidence is score for True or its complement for Fake, not factual
        certainty.
      required:
      - label
      - confidence
      - writing_score
      properties:
        label:
          type: string
          enum:
          - Fake
          - true
        confidence:
          type: number
          minimum: 0
          maximum: 1
        writing_score:
          type: number
          minimum: 0
          maximum: 1
      additionalProperties: false
...
    ReservedCriterion:
      type: object
      required:
      - available
      - score
      - status
      - planned_model
      - planned_flow
      properties:
        available:
          type: boolean
          const: false
        score:
          type: 'null'
        status:
          type: string
          const: NOT_IMPLEMENTED
        planned_model:
          type: string
          const: Ashg2099/xlm-roberta-factchecker
        planned_flow:
          type: array
          items:
            type: string
      additionalProperties: false
    FinalScore:
      type: object
      required:
      - score
      - coverage
      - intended_weights
      - effective_weights
      - formula
      - limitation
      properties:
        score:
          type:
          - number
          - 'null'
          minimum: 0
          maximum: 100
        coverage:
          type: number
          enum:
          - 0
          - 40
          - 60
          - 100
        intended_weights:
          type: object
          required:
          - verifiable_facts
          - writing_style
          properties:
            verifiable_facts:
              type: number
              const: 0.6
            writing_style:
              type: number
              const: 0.4
          additionalProperties: false
        effective_weights:
          type: object
          additionalProperties:
            type: number
        formula:
          type: string
        limitation:
          type: string
      additionalProperties: false
    PipelineVersion:
      type: object
      required:
      - id
      - rules_version
      - fact_check_mapping_version
      - writing_model_name
      - dependency_versions
      properties:
        id:
          type: string
        rules_version:
          type: string
        fact_check_mapping_version:
          type: string
        writing_model_name:
          type: string
        writing_model_revision:
          type: string
        dependency_versions:
          type: object
      additionalProperties: false
    ErrorInfo:
      type: object
      required:
      - code
      - message
      - retryable
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean
        details:
          type: object
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
      - status
      - error
      properties:
        status:
          $ref: '#/components/schemas/AnalysisStatus'
        error:
          $ref: '#/components/schemas/ErrorInfo'
      additionalProperties: false
    CriterionStatus:
      enum:
      - EXECUTED
      - UNAVAILABLE
      - ERROR
      - NOT_IMPLEMENTED
      title: CriterionStatus
      type: string